requirements.txt
pytest>=7.4.0
requests>=2.31.0
allure-pytest>=2.13.2
PyYAML>=6.0
jsonpath>=0.82
loguru>=0.7.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0
